third_party/blink/web_tests/fast/dom/subtree-modified-attributes-expected.txt

DOMSubtreeModified should fire when attributes are added or removed, but not modified (see bug 81141)

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS expected is true
PASS expected is true
PASS expected is true
PASS expected is true
PASS successfullyParsed is true

TEST COMPLETE
